Refactor(#126): 반응형 Masonry 레이아웃 리팩토링
