+++ components/card/MasonryCard/CardImage.tsx
@@ -1,0 +1,310 @@
+import { Box, Flex, Text } from '@chakra-ui/react';
+import { FaEye, FaThumbsUp } from 'react-icons/fa';
+import { FaArrowRightLong } from 'react-icons/fa6';
+
+import { formatArtistValue } from '@/hook/useFormatArtistValue';
+import { useModifiedImageUrl } from '@/hook/useModifiedImageUrl';
+import styles from '@/styles/MasonryCard.module.scss';
+import { useResponsive } from '@/hook/useResponsive';
+import Image from 'next/image';
+import Link from 'next/link';
+
+type Props = {
+  data: ArtworkList | GalleryArtworkList;
+};
+
+export default function CardImage({ data }: Props) {
+  const {
+    img_url,
+    img_url_list,
+    title,
+    board,
+    category,
+    view,
+    like,
+    comment,
+    date,
+    deleted,
+    id,
+    url,
+  } = data;
+  const isMobile = useResponsive();
+  const modifiedUrl300 = useModifiedImageUrl({
+    url: img_url_list[0],
+    size: 300,
+  });
+
+  if (isMobile) {
+    return (
+      <Link href={`/artwork/${id}`}>
+        <Image
+          width={236}
+          height={236}
+          src={
+            img_url === ''
+              ? 'http://via.placeholder.com/236x236'
+              : modifiedUrl300
+          }
+          alt={title}
+          style={{
+            maxHeight: '430px',
+            objectFit: 'cover',
+            borderRadius: '20px',
+            filter: deleted ? 'blur(6px)' : 'none', // 블러 처리
+            background: 'rgb(245, 245, 245)',
+          }}
+          unoptimized
+        />
+      </Link>
+    );
+  }
+
+  return (
+    <Box className={styles.cardContainer} position="relative">
+      <Image
+        width={236}
+        height={236}
+        src={
+          img_url === '' ? 'http://via.placeholder.com/236x236' : modifiedUrl300
+        }
+        alt={title}
+        style={{
+          maxHeight: '430px',
+          objectFit: 'cover',
+          borderRadius: '20px',
+          filter: deleted ? 'blur(6px)' : 'none', // 블러 처리
+          background: 'rgb(245, 245, 245)',
+        }}
+        unoptimized
+      />
+      <Box
+        className={styles.overlay}
+        position="absolute"
+        top={0}
+        right={0}
+        bottom={0}
+        left={0}
+        borderRadius="20px"
+        zIndex={1}
+      />
+      <Flex
+        className={styles.hoverContainer}
+        // style={
+        //   imageHeight && imageHeight <= 200
+        //     ? {
+        //         padding: 0,
+        //       }
+        //     : undefined
+        // }
+        flexDir="column"
+        position="absolute"
+        top={0}
+        right={0}
+        bottom={0}
+        left={0}
+        borderRadius="20px"
+        border="1px solid rgba(255, 255, 255, 0.70)"
+        zIndex={2}
+        background="rgba(0, 0, 0, 0.5)"
+        justifyContent="space-between"
+        alignItems="center"
+        color="white"
+        fontSize="2rem"
+        fontWeight="bold"
+        // cursor="pointer"
+        p={['0.5rem 0', '1rem 0']}
+      >
+        <Box
+          className={styles.body}
+          width="100%"
+          display="flex"
+          flexDir="column"
+          justifyContent="flex-end"
+          alignItems="center"
+          h="80%"
+          borderBottom="1px solid rgba(255, 255, 255, 0.70)"
+          p="1rem"
+        >
+          <Text
+            className={styles.board}
+            w="100%"
+            textAlign="left"
+            fontSize={['sm', 'lg']}
+            fontWeight="600"
+            whiteSpace="nowrap"
+            overflow="hidden"
+            textOverflow="ellipsis"
+          >
+            {board.replace(/&#\d+;/g, '').trim()}
+          </Text>
+          <Flex
+            className={styles.info}
+            flexDir="row"
+            w="100%"
+            justifyContent="flex-start"
+            alignItems="center"
+            gap="0.5rem"
+          >
+            <Box
+              display="flex"
+              flexDir="row"
+              justifyContent="center"
+              alignItems="center"
+              gap="0.3rem"
+              whiteSpace="nowrap"
+              overflow="hidden"
+              textOverflow="ellipsis"
+              maxWidth="100%"
+            >
+              <Box w="14px" h="14px">
+                <svg
+                  xmlns="http://www.w3.org/2000/svg"
+                  fill="none"
+                  viewBox="0 0 24 24"
+                  strokeWidth="1.5"
+                  stroke="currentColor"
+                  className="w-6 h-6"
+                >
+                  <path
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    d="M12 6v6h4.5m4.5 0a9 9 0 11-18 0 9 9 0 0118 0z"
+                  />
+                </svg>
+              </Box>
+              <Text
+                fontSize={['xs', 'sm', '15px']}
+                fontWeight="400"
+                color="#FFFFFFB3"
+                textAlign="center"
+              >
+                {date?.split(' ')[0].slice(2, -1)}
+              </Text>
+            </Box>
+            <Box
+              display="flex"
+              flexDir="row"
+              justifyContent="center"
+              alignItems="center"
+              gap="0.3rem"
+            >
+              <FaEye
+                size="14px"
+                // style={widthValue === '180px' ? iconStyleMobile : iconStyle}
+                color="#FFFFFFB3"
+              />
+              <Text
+                fontSize={['xs', 'sm', '15px']}
+                fontWeight="400"
+                color="#FFFFFFB3"
+              >
+                {formatArtistValue(view)}
+              </Text>
+            </Box>
+            <Box
+              display="flex"
+              flexDir="row"
+              justifyContent="center"
+              alignItems="center"
+              gap="0.3rem"
+            >
+              <FaThumbsUp
+                size="14px"
+                // style={widthValue === '180px' ? iconStyleMobile : iconStyle}
+                color="#FFFFFFB3"
+              />
+              <Text
+                fontSize={['xs', 'sm', '15px']}
+                fontWeight="400"
+                color="#FFFFFFB3"
+              >
+                {formatArtistValue(like)}
+              </Text>
+            </Box>
+          </Flex>
+        </Box>
+        <Box
+          className={styles.footer}
+          display="flex"
+          flexDir="row"
+          gap="0.5rem"
+          justifyContent="center"
+          alignItems="center"
+          w="100%"
+          h="20%"
+          pt="1rem"
+          px="1rem"
+        >
+          {/* <NextLink
+              className={styles.btns}
+              href={`/artists/${authorName}`}
+              style={{
+                flex: 0.5,
+              }}
+              prefetch={false}
+            >
+              <Box
+                className={styles.textBox}
+                w="100%"
+                display="flex"
+                flexDir="row"
+                alignItems="center"
+                justifyContent="center"
+                gap="4px"
+                _hover={{
+                  textDecoration: 'none',
+                  cursor: 'pointer',
+                  backgroundColor: 'black',
+                  color: 'white',
+                }}
+                // rel="noopener noreferrer" // 보안상의 이유료 이 부분도 추가합니다.
+                borderRadius="800px"
+                background="white"
+                padding="8px 10px"
+                fontSize="xs"
+                color="black"
+                transition="all 0.2s ease-in-out"
+              >
+                <MdPerson />
+                작가
+              </Box>
+            </NextLink> */}
+          <Link
+            className={styles.btns}
+            href={`/artwork/${id}`}
+            style={{
+              flex: 1,
+            }}
+          >
+            <Box
+              className={styles.textBox}
+              w="100%"
+              display="flex"
+              flexDir="row"
+              alignItems="center"
+              justifyContent="center"
+              gap="5px"
+              _hover={{
+                textDecoration: 'none',
+                cursor: 'pointer',
+                background: 'linear-gradient(92deg, #ff58a2 0%, #ff93c3 100%)',
+              }}
+              // rel="noopener noreferrer" // 보안상의 이유료 이 부분도 추가합니다.
+              borderRadius="11px"
+              background="linear-gradient(92deg, #FF4195 0%, #FF72B0 100%)"
+              padding="8px 10px"
+              fontSize="xs"
+              fontWeight="600"
+              color="white"
+              transition="all 0.2s ease-in-out"
+            >
+              {/* 왁물원<span>에서 보기</span> */}
+              자세히보기
+              <FaArrowRightLong />
+            </Box>
+          </Link>
+        </Box>
+      </Flex>
+    </Box>
+  );
+}
