+++ components/views/MasonryView.tsx
@@ -1,5 +1,5 @@
-import { Box, useBreakpointValue } from '@chakra-ui/react';
-import Masonry from 'react-masonry-css';
+import Measure from 'react-measure';
+import Masonry, { ResponsiveMasonry } from 'react-responsive-masonry';
 
 import MasonryCard from '@/components/card/MasonryCard';
 
@@ -9,46 +9,53 @@ type Props = {
 };
 
 export default function MasonryView({ artworks, isDeletedVisible }: Props) {
-  // const article_link = useResponsiveLink('', 'article');
-  const widthValue = useBreakpointValue({ base: '180px', sm: '236px' });
-  const widthValue2 = useBreakpointValue({ base: '188px', sm: '252px' });
-  const breakpointColumnsObj = {
-    default: 7,
-    1792: 6,
-    1528: 5,
-    1290: 4,
-    1036: 3,
-    792: 2,
-    556: 1,
-  };
-
   const content = () => {
     if (isDeletedVisible) {
-      return artworks.map((artwork, index) => (
-        <MasonryCard key={index} artwork={artwork} />
+      return artworks.map((artwork) => (
+        <Measure key={artwork.id}>
+          {({ measureRef }) => (
+            <div ref={measureRef}>
+              <MasonryCard key={artwork.id} artwork={artwork} />
+            </div>
+          )}
+        </Measure>
       ));
     }
 
     if (!isDeletedVisible) {
-      return artworks.map((artwork, index) =>
-        !artwork.deleted ? <MasonryCard key={index} artwork={artwork} /> : null
+      return artworks.map((artwork) =>
+        !artwork.deleted ? (
+          <Measure key={artwork.id}>
+            {({ measureRef }) => (
+              <div
+                ref={measureRef}
+                style={{
+                  display: 'flex',
+                  justifyContent: 'center',
+                  alignItems: 'center',
+                }}
+              >
+                <MasonryCard key={artwork.id} artwork={artwork} />
+              </div>
+            )}
+          </Measure>
+        ) : null
       );
     }
   };
 
   return (
-    <Box w="100%" mx="auto" position="relative">
-      <Box
-        as={Masonry}
-        breakpointCols={breakpointColumnsObj}
-        className="my-masonry-grid"
-        columnClassName="my-masonry-grid_column"
-        display="flex"
-        justifyContent="center"
-        width="100%"
-      >
-        {content()}
-      </Box>
-    </Box>
+    <ResponsiveMasonry
+      columnsCountBreakPoints={{
+        300: 2,
+        756: 3,
+        1008: 4,
+        1260: 5,
+        1528: 6,
+        1792: 7,
+      }}
+    >
+      <Masonry gutter="10px">{content()}</Masonry>
+    </ResponsiveMasonry>
   );
 }
