+++ package.json
@@ -38,7 +38,8 @@
     "react-hot-toast": "^2.4.1",
     "react-icons": "^5.0.1",
     "react-intersection-observer": "^9.8.1",
-    "react-masonry-css": "^1.0.16",
+    "react-measure": "^2.5.2",
+    "react-responsive-masonry": "^2.2.0",
     "react-snowfall": "^1.2.1",
     "react-spinners": "^0.13.8",
     "react-use": "^17.5.0",
@@ -52,6 +53,8 @@
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "@types/react-measure": "^2.0.12",
+    "@types/react-responsive-masonry": "^2.1.3",
     "@types/uuid": "^9.0.8",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "@typescript-eslint/parser": "^6.21.0",
